requirements.txt
python-telegram-bot==20.0a1
python-dotenv==0.21
google-generativeai==0.7.2